make jei categories client only
