+++ overrides/groovy/classes/content/jei/category/GenericRecipeCategory.groovy
@@ -1,3 +1,4 @@
+// side: client
 package classes.content.jei.category
 
 import mezz.jei.api.gui.IDrawable
