+++ overrides/groovy/classes/content/jei/category/MatrixStabilizerCategory.groovy
@@ -1,3 +1,4 @@
+// side: client
 package classes.content.jei.category
 
 import classes.content.thaumcraft.MatrixStabilityStorage
